Add Sample Game Development Page
This commit adds a sample html and css file for the game development and programming page. The style itself looks solid. The only thing that needs to be done is hooking up the "Play Game" buttons to the actual page (and then customizing a typical Unity WebGL page).

Note to Self: Ask ChatGPT to edit each project to include a .gif or .png file in its data section. Also, tell it to abandon the idea of a shared css file. It broke the landing page.